manage listings

+++ app/Http/Controllers/ListingController.php
@@ -111,6 +111,12 @@ class ListingController extends Controller
      */
     public function update(Request $request, Listing $listing)
     {
+
+        // Make sure logged in user is owner
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => 'required',
@@ -141,7 +147,23 @@ class ListingController extends Controller
      */
     public function destroy(Listing $listing)
     {
+        // Make sure logged in user is owner
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully');
+    }
+
+    /**
+     * Manage Listings
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function manage()
+    {
+        return view('listings/manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
